fix: tidy up and move files around
