+++ requirements.txt
@@ -1,6 +1,23 @@
-./etebase_server
 Django==3.1.4
-django-cors-headers==3.7.0
-django-ninja==0.12.2
-psycopg2-binary==2.8.6
 uvicorn==0.13.2
+fastapi~=0.63.0
+starlette~=0.13.6
+h11~=0.11.0
+pip~=21.0.1
+six~=1.15.0
+pytz~=2020.4
+PyYAML~=5.3.1
+sqlparse~=0.4.1
+asgiref~=3.3.1
+click~=7.1.2
+uvloop~=0.14.0
+pydantic~=1.7.3
+websockets~=8.1
+hiredis~=1.1.0
+httptools~=0.1.1
+watchgod~=0.6
+aioredis~=1.3.1
+pycparser~=2.20
+aiofiles~=0.6.0
+setuptools~=54.2.0
+msgpack~=1.0.2
