+++ server/settings/base.py
@@ -13,7 +13,7 @@ def get_env_value(env_variable):
     try:
         return os.environ[env_variable]
     except KeyError:
-        error_msg = 'Set the {} environment variable'.format(env_variable)
+        error_msg = "Set the {} environment variable".format(env_variable)
         raise ImproperlyConfigured(error_msg)
 
 
@@ -34,7 +34,9 @@ CORS_ALLOW_CREDENTIALS = True
 DATABASES = {
     "default": {
         "ENGINE": "django.db.backends.sqlite3",
-        "NAME": os.environ.get("ETEBASE_DB_PATH", os.path.join(BASE_DIR, "../", "db.sqlite3")),
+        "NAME": os.environ.get(
+            "ETEBASE_DB_PATH", os.path.join(BASE_DIR, "../", "db.sqlite3")
+        ),
     }
 }
